pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PySigmaKoki"
version = "2.1.8"
description = "Python Interface for Instruments by Sigma Koki"
authors = [
    { name = "Akira Okumura", email = "oxon@mac.com" }
]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "BSD License"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent"
]
dependencies = [
    "pyserial"
]

[project.urls]
Repository = "https://github.com/akira-okumura/PySigmaKoki.git"
Issues = "https://github.com/akira-okumura/PySigmaKoki/issues"
